requirement.txt
asgiref==3.2.5
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
chardet==3.0.4
cryptography==2.8
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.4
django-ajax==0.3.0
django-allauth==0.41.0
django-appconf==1.0.4
django-autoslug==1.9.6
django-backupdb==0.7.0
django-ckeditor==5.9.0
django-countries==6.1.2
django-decouple==2.1
django-hitcount==1.3.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
oauthlib==3.1.0
pilkit==2.0
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-decouple==3.4
python3-openid==3.1.0
pytz==2019.3
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
sqlparse==0.3.1
stripe==2.46.0
tinycss2==1.0.2
Unidecode==1.1.1
urllib3==1.25.8
WeasyPrint==51
webencodings==0.5.1
xhtml2pdf==0.2.4
